refactor: 현장톡 Activity 패키지 이동 및 코드 수정
