+++ android/app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:theme="@style/Theme.YaguBogu"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.livetalk.LivetalkChatActivity"
+            android:name=".presentation.livetalk.chat.LivetalkChatActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustResize" />
         <activity
